Fix safari iframe issues and tweak styles
