+++ widget/index.html
@@ -1,5 +1,11 @@
 <html>
   <head>
+    <meta charset="utf-8" />
+    <meta http-equiv="x-ua-compatible" content="ie=edge" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1, shrink-to-fit=no"
+    />
     <title>Changelog Widget</title>
     <script src="./widget.js" defer></script>
   </head>
