+++ widget/src/styles.css
@@ -1,4 +1,4 @@
-.changelogNotification {
+.notification {
   display: inline-block;
   border-radius: 50%;
   width: 22px;
@@ -17,31 +17,31 @@
   animation: growIn 500ms cubic-bezier(0.175, 0.985, 0.1, 1.035);
 }
 
-.changelogIframe {
+.iframe {
   display: block;
-  position: absolute;
-  height: 100%;
-  width: 100%;
+  position: fixed;
+  height: 100vh;
+  width: 100vw;
   top: 0;
   right: 0;
   margin: 0;
   padding: 0;
   border: none;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
   transform: translateX(100%);
   animation: slideOut 500ms cubic-bezier(0.175, 0.985, 0.1, 1.035);
 }
 
-.changelogIframeOpen {
+.iframeOpen {
   transform: translateX(0);
-  width: 100%;
   animation: slideIn 500ms cubic-bezier(0.175, 0.985, 0.1, 1.035);
 }
 
-.changelogIframeHidden {
+.iframeHidden {
   display: none;
 }
 
-.changelogOverlay {
+.overlay {
   height: 100%;
   width: 100%;
   background: black;
@@ -52,22 +52,18 @@
   transition: opacity 350ms linear;
 }
 
-.changelogOverlayOpen {
-  visibility: hidden;
+.overlayOpen {
+  opacity: 0.5;
+  visibility: visible;
 }
 
-.changelogOverlayHidden {
+.overlayHidden {
   visibility: hidden;
 }
 
 @media (min-width: 800px) {
-  .changelogIframe {
-    width: 600px;
-  }
-
-  .changelogOverlayOpen {
-    opacity: 0.5;
-    visibility: visible;
+  .iframe {
+    width: 399px;
   }
 }
